split skeleton debug to public and internal methods (#434)
