+++ SkeletonView.xcodeproj/xcshareddata/IDETemplateMacros.plist
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>FILEHEADER</key>
+	<string>
+//  Copyright SkeletonView. All Rights Reserved.
+//
+//  Licensed under the MIT License (the &quot;License&quot;);
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//      https://opensource.org/licenses/MIT
+//
+//  ___FILENAME___
+//
+//  Created by ___FULLUSERNAME___ on ___DATE___.</string>
+</dict>
+</plist>
